remove old view

+++ app/src/main/java/com/jackq/funfurniture/AR/util/Utilities.java
@@ -178,22 +178,22 @@ public class Utilities
 
         if (cameraX != null && cameraX.length > 0)
         {
-            cameraX[0] = (int) ((scaledUpX / (float) scaledUpVideoWidth) * videoWidth);
+            cameraX[0] = (int) ((scaledUpX / scaledUpVideoWidth) * videoWidth);
         }
 
         if (cameraY != null && cameraY.length > 0)
         {
-            cameraY[0] = (int) ((scaledUpY / (float) scaledUpVideoHeight) * videoHeight);
+            cameraY[0] = (int) ((scaledUpY / scaledUpVideoHeight) * videoHeight);
         }
 
         if (cameraDX != null && cameraDX.length > 0)
         {
-            cameraDX[0] = (int) (((float) screenDX / (float) scaledUpVideoWidth) * videoWidth);
+            cameraDX[0] = (int) (((float) screenDX / scaledUpVideoWidth) * videoWidth);
         }
 
         if (cameraDY != null && cameraDY.length > 0)
         {
-            cameraDY[0] = (int) (((float) screenDY / (float) scaledUpVideoHeight) * videoHeight);
+            cameraDY[0] = (int) (((float) screenDY / scaledUpVideoHeight) * videoHeight);
         }
     }
